@@ -14,15 +14,15 @@ description: A Social app UI made with flutter.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.2
-  google_fonts: ^1.0.0
-  flutter_staggered_animations: ^0.1.2
-  lottie: ^0.6.0
+  cupertino_icons: ^1.0.5
+  google_fonts: ^4.0.4
+  flutter_staggered_animations: ^1.1.1
+  lottie: ^2.4.0
 
 dev_dependencies:
   flutter_test:
