@@ -11,37 +11,59 @@ class ThemeConfig {
   static Color badgeColor = Colors.red;
 
   static ThemeData lightTheme = ThemeData(
-    backgroundColor: lightBG,
     primaryColor: lightPrimary,
-    accentColor: lightAccent,
     scaffoldBackgroundColor: lightBG,
     appBarTheme: AppBarTheme(
+      backgroundColor: lightBG,
       elevation: 0,
-      textTheme: TextTheme(
-        headline6: TextStyle(
+      toolbarTextStyle: TextTheme(
+        titleLarge: TextStyle(
           color: darkBG,
           fontSize: 18.0,
           fontWeight: FontWeight.w800,
         ),
-      ),
+      ).bodyMedium,
+      titleTextStyle: TextTheme(
+        titleLarge: TextStyle(
+          color: darkBG,
+          fontSize: 18.0,
+          fontWeight: FontWeight.w800,
+        ),
+      ).titleLarge,
+    ),
+    colorScheme: ColorScheme.fromSwatch().copyWith(
+      secondary: lightAccent,
+      background: lightBG,
+      brightness: Brightness.light,
     ),
   );
 
   static ThemeData darkTheme = ThemeData(
     brightness: Brightness.dark,
-    backgroundColor: darkBG,
     primaryColor: darkPrimary,
-    accentColor: darkAccent,
     scaffoldBackgroundColor: darkBG,
     appBarTheme: AppBarTheme(
+      backgroundColor: darkBG,
       elevation: 0,
-      textTheme: TextTheme(
-        headline6: TextStyle(
+      toolbarTextStyle: TextTheme(
+        titleLarge: TextStyle(
           color: lightBG,
           fontSize: 18.0,
           fontWeight: FontWeight.w800,
         ),
-      ),
+      ).bodyMedium,
+      titleTextStyle: TextTheme(
+        titleLarge: TextStyle(
+          color: lightBG,
+          fontSize: 18.0,
+          fontWeight: FontWeight.w800,
+        ),
+      ).titleLarge,
+    ),
+    colorScheme: ColorScheme.fromSwatch().copyWith(
+      secondary: darkAccent,
+      background: darkBG,
+      brightness: Brightness.dark,
     ),
   );
 }
