@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 
 class CustomButton extends StatelessWidget {
   final String label;
-  final Function onPressed;
-  final Color color;
+  final Function? onPressed;
+  final Color? color;
 
   CustomButton({
     this.label = 'Continue',
@@ -12,20 +12,34 @@ class CustomButton extends StatelessWidget {
   });
 
   bool isSmallScreen(BuildContext context) {
-    return MediaQuery.of(context).size.width < 800;
+    return MediaQuery
+        .of(context)
+        .size
+        .width < 800;
   }
 
   @override
   Widget build(BuildContext context) {
     return Container(
       height: 50.0,
-      width: MediaQuery.of(context).size.width,
-      child: FlatButton(
-        onPressed: onPressed,
-        color: color ?? Theme.of(context).accentColor,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.all(
-            Radius.circular(10.0),
+      width: MediaQuery
+          .of(context)
+          .size
+          .width,
+      child: TextButton(
+        onPressed: () => onPressed!(),
+        style: ButtonStyle(
+          backgroundColor: MaterialStateProperty.all(
+              color ?? Theme
+                  .of(context)
+                  .colorScheme
+                  .secondary),
+          shape: MaterialStateProperty.all(
+            RoundedRectangleBorder(
+              borderRadius: BorderRadius.all(
+                Radius.circular(10.0),
+              ),
+            ),
           ),
         ),
         child: Text(
