@@ -1,14 +1,14 @@
 class Validations {
-  static String validateName(String value) {
-    if (value.isEmpty) return 'Name is Required.';
+  static String? validateName(String? value) {
+    if (value == null || value.isEmpty) return 'Name is Required.';
     final RegExp nameExp = new RegExp(r'^[A-za-zğüşöçİĞÜŞÖÇ ]+$');
     if (!nameExp.hasMatch(value))
       return 'Please enter only alphabetical characters.';
     return null;
   }
 
-  static String validateEmail(String value) {
-    if (value.isEmpty) return 'Please enter an Email Address.';
+  static String? validateEmail(String? value) {
+    if (value == null || value.isEmpty) return 'Please enter an Email Address.';
     final RegExp nameExp = new RegExp(
         r"^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,2"
         r"53}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,253}[a-z"
@@ -17,8 +17,9 @@ class Validations {
     return null;
   }
 
-  static String validatePassword(String value) {
-    if (value.isEmpty || value.length < 6) return 'Please enter a valid password.';
+  static String? validatePassword(String? value) {
+    if (value == null || value.isEmpty || value.length < 6)
+      return 'Please enter a valid password.';
     return null;
   }
 }
